Updated for PHP 7 compatibility
