+++ src/CLIResultPrinter.php
@@ -59,7 +59,7 @@ class CLIResultPrinter implements ResultPrinterInterface
         }
 
         foreach ($context->getErrors() as $error) {
-            $this->output->writeln($error->getMessage());
+            $this->output->writeln($error->getText());
         }
 
         $this->output->writeln('');
