+++ src/CompatibilityViolation/Message.php
@@ -3,64 +3,24 @@
 namespace Sstalle\php7cc\CompatibilityViolation;
 
 use PhpParser\Node;
+use Sstalle\php7cc\AbstractBaseMessage;
 
-class Message
+class Message extends AbstractBaseMessage
 {
-    /**
-     * @var string
-     */
-    protected $rawText;
-
-    /**
-     * @var string
-     */
-    protected $text;
-
-    /**
-     * @var int
-     */
-    protected $line;
-
     /**
      * @var Node[]
      */
     protected $nodes;
 
     /**
-     * @param string $text
-     * @param int    $line
-     * @param Node[] $nodes
+     * @param string   $text
+     * @param int|null $line
+     * @param Node[]   $nodes
      */
-    public function __construct($text, $line, array $nodes)
+    public function __construct($text, $line = null, array $nodes = array())
     {
-        $this->rawText = $text;
-        $this->line = $line;
+        parent::__construct($text, $line);
         $this->nodes = $nodes;
-        $this->text = $this->generateText();
-    }
-
-    /**
-     * @return string
-     */
-    public function getRawText()
-    {
-        return $this->rawText;
-    }
-
-    /**
-     * @return string
-     */
-    public function getText()
-    {
-        return $this->text;
-    }
-
-    /**
-     * @return int
-     */
-    public function getLine()
-    {
-        return $this->line;
     }
 
     /**
