+++ test/resource/escapedUnicodeCodepoint/php5/escapedUnicodeCodepointInHeredoc.test
@@ -1,0 +1,25 @@
+Escaped unicode codepoints in heredoc
+PHP <7.0
+-----
+<?php
+$str = <<<EOD
+    \u{xyz}
+EOD;
+
+-----
+Unicode codepoint escaping "\u{xyz}" in a string
+-----
+<?php
+$str = <<<EOD
+    \\u{xyz}
+EOD;
+
+-----
+
+-----
+<?php
+$str = <<<EOD
+    \u202e
+EOD;
+
+-----
