+++ test/resource/escapedUnicodeCodepoint/php5/escapedUnicodeCodepointInString.test
@@ -1,4 +1,5 @@
 Escaped unicode codepoints in double quoted string
+PHP <7.0
 -----
 <?php $str = "\u{xyz}";
 -----
