+++ test/resource/escapedUnicodeCodepoint/php7/escapedUnicodeCodepointInHeredoc.test
@@ -1,4 +1,5 @@
 Escaped unicode codepoints in heredoc
+PHP >=7.0
 -----
 <?php
 $str = <<<EOD
@@ -6,7 +7,7 @@ $str = <<<EOD
 EOD;
 
 -----
-Unicode codepoint escaping "\u{xyz}" in a string
+Invalid UTF-8 codepoint escape sequence
 -----
 <?php
 $str = <<<EOD
