+++ test/resource/invalidOctalLiteral/php5/invalidOctalLiteral.test
@@ -1,4 +1,5 @@
 Invalid octal literal
+PHP <7.0
 -----
 <?php
 $i = 0781;
